+++ lib/Service/AccountService.php
@@ -27,7 +27,7 @@ use OCA\Mail\Account;
 use OCA\Mail\Db\MailAccount;
 use OCA\Mail\Db\MailAccountMapper;
 use OCA\Mail\Exception\ServiceException;
-use OCA\Mail\Service\DefaultAccount\DefaultAccountManager;
+use OCA\Mail\Service\DefaultAccount\Manager as DefaultAccountManager;
 use OCP\IL10N;
 
 class AccountService {
